move pnpm-project test utils to testtools package and use from there for consistency and added reuse possibilities.
removed pnpm store initialization from the scaffolding process for now, because it's slightly easier dependency wise and because doing this in scaffold is something I want to revisit anyway. I'd like to end up in a system where we can include a large superset of the dependencies in play. Not sure if that's attainable at this point.

Given that we now don't run pnpm store initialization in test-scaffold, this means the performance hit will happen on (multiple?) first tests. This is not acceptable for the end result, but as an intermediate state in this PR it's good enough.

Note as well that the exact implementation of what is now pnpm-project.ts changed a bit: it now uses the tmp package for consistency, optionally includes some standard dependencies, and includes workspace overrides (because the pnpm behavior with regards to deep dependencies is different fron npm).